refactored server logic and candidate logic on the server side
